+++ packages/pic/src/pocket-ic-server.ts
@@ -64,13 +64,16 @@ export class PocketIcServer {
     options: StartServerOptions = {},
   ): Promise<PocketIcServer> {
     const binPath = this.getBinPath();
+
+    console.log(binPath);
+
     await this.assertBinExists(binPath);
 
     const pid = process.ppid;
     const picFilePrefix = `pocket_ic_${pid}`;
     const portFilePath = tmpFile(`${picFilePrefix}.port`);
 
-    const serverProcess = spawn(binPath, ['--pid', pid.toString()]);
+    const serverProcess = spawn(binPath, ['--port-file', portFilePath]);
 
     if (options.showRuntimeLogs) {
       serverProcess.stdout.pipe(process.stdout);